Prepare for ember-cli-fastboot 1.0
