--- app/instance-initializers/fastboot/algolia-config.js
@@ -1,17 +1,0 @@
-import Ember from 'ember';
-import config from '../../config/environment';
-
-const { set } = Ember;
-
-export function initialize( appInstance ) {
-  const service = appInstance.lookup('service:algolia');
-  let algoliaConfig = config['ember-algolia'];
-
-  set(service, 'client', algoliasearch(algoliaConfig.algoliaId, algoliaConfig.algoliaKey));
-
-}
-
-export default {
-  name: 'algolia-config',
-  initialize
-};
