+++ package.json
@@ -21,13 +21,12 @@
   },
   "dependencies": {
     "broccoli-funnel": "^1.1.0",
-    "ember-cli-babel": "^5.1.7",
-    "fastboot-filter-initializers": "0.0.2"
+    "ember-cli-babel": "^5.1.7"
   },
   "devDependencies": {
     "broccoli-asset-rev": "^2.4.5",
     "ember-ajax": "^2.4.1",
-    "ember-cli": "2.12.0-beta.1",
+    "ember-cli": "2.14.0-beta.2",
     "ember-cli-dependency-checker": "^1.3.0",
     "ember-cli-eslint": "^3.0.3",
     "ember-cli-htmlbars": "^1.1.1",
