+++ public/algoliasearch.js
